+++ app/src/main/java/com/example/user/musicpal/model/dao/Service.java
@@ -32,5 +32,5 @@ public interface Service {
     Call<ContenedorDeCanciones> obtenerCancionesPorArtistaConId(@Path("id") int id);
 
     @GET("playlist/{id}/tracks")
-    Call<ContenedorDeCanciones> obtenerCancionesPorPlaylistConId(@Path("id")int id);
+    Call<ContenedorDeCanciones> obtenerCancionesPorPlaylistConId(@Path("id")double id);
 }
